Added Buttons to create new notes.
Using jquery 'append' I've created some buttons that allow us to add new
notes and pins. The problem is append will always add the new notes or
pins to the end of the 'main' div.  Jquery 'after' or 'insertafter' may
be a better choice, but append will substitute for now.